Поправил класс Actions

+++ modules/data_classes/data_classes.py
@@ -3,10 +3,10 @@ from dataclasses import dataclass
 
 @dataclass
 class Object_position:
-    x1: float  # Лево вверх, координаты X
-    y1: float  # Лево вверх, координаты Y
-    x2: float  # Право низ, координаты X
-    y2: float  # Право низ, координаты Y
+    x1: int  # Лево вверх, координаты X
+    y1: int  # Лево вверх, координаты Y
+    x2: int  # Право низ, координаты X
+    y2: int  # Право низ, координаты Y
 
     def top(self):
         return self.y2 - self.y1
